2022-04-24 leave some instruction fun not implemented

+++ src/add.c
@@ -1,12 +1,12 @@
 ﻿#include<stdio.h>
 #include<stdint.h>
-unsigned add(unsigned i1,unsigned i2){
+uint64_t add(uint64_t i1,uint64_t i2){
     return i1+i2;
 }
 
 int main(){
-    int i1 = 2;
-    int i2 = 5;
+    uint64_t i1 = 2;
+    uint64_t i2 = 5;
     add(i1,i2);
     return 0;
 }
